move pools to stateful namespace
