+++ include/hegel/generators/collections.h
@@ -48,36 +48,6 @@ namespace hegel::generators {
                                         ///< sizes)
     };
 
-    template <typename T> class VariablesGenerator;
-
-    template <typename T> class Pool {
-      public:
-        explicit Pool(const TestCase& tc) : tc_(tc) {
-            pool_id_ = hegel::internal::new_pool(tc);
-        }
-
-        void add(const T& element) {
-            int64_t var_id = hegel::internal::pool_add(tc_, pool_id_);
-            pool_.emplace(var_id, element);
-        }
-
-        void add(T&& element) {
-            int64_t var_id = hegel::internal::pool_add(tc_, pool_id_);
-            pool_.emplace(var_id, std::move(element));
-        }
-
-        std::size_t size() const { return pool_.size(); }
-
-        Pool(const Pool&) = delete;
-
-      private:
-        const TestCase& tc_;
-        int64_t pool_id_;
-        std::map<int64_t, T> pool_;
-
-        friend class VariablesGenerator<T>;
-    };
-
     /// @cond INTERNAL
     namespace detail {
 
@@ -402,44 +372,6 @@ namespace hegel::generators {
     Generator<std::array<T, N>> arrays(Generator<T> element) {
         return Generator<std::array<T, N>>(
             new ArrayGenerator<T, N>(std::move(element)));
-    }
-
-    template <typename T> class VariablesGenerator : public IGenerator<T> {
-      public:
-        explicit VariablesGenerator(Pool<T>& p, bool consume)
-            : p_(p), consume_(consume) {}
-
-        T do_draw(const TestCase& tc) const override {
-            int64_t variable =
-                hegel::internal::draw_variable(tc, p_.pool_id_, consume_);
-
-            auto it = p_.pool_.find(variable);
-            // GCOVR_EXCL_START
-            if (it == p_.pool_.end()) {
-                throw std::runtime_error(
-                    "Pool state diverged between the engine and the "
-                    "client, or a bug in the pool bookkeeping.");
-            }
-            // GCOVR_EXCL_STOP
-            if (consume_) {
-                auto val = std::move(it->second);
-                p_.pool_.erase(it);
-                return val;
-            }
-            return it->second;
-        }
-
-      private:
-        Pool<T>& p_;
-        bool consume_;
-    };
-
-    template <typename T> Generator<T> values_consumed(Pool<T>& p) {
-        return Generator<T>(new VariablesGenerator<T>(p, true));
-    }
-
-    template <typename T> Generator<T> values_reusable(Pool<T>& p) {
-        return Generator<T>(new VariablesGenerator<T>(p, false));
     }
 
     /// @}
