+++ tests/test_generators_collections.cpp
@@ -10,43 +10,6 @@
 
 namespace gs = hegel::generators;
 
-TEST(Pools, PoolsRoundTrip) {
-    hegel::test([](hegel::TestCase& tc) {
-        gs::Pool<int> pool = gs::Pool<int>(tc);
-        std::set<int> original_set =
-            tc.draw(gs::sets(gs::integers<int>(), {.max_size = 10}));
-
-        for (int num : original_set) {
-            pool.add(num);
-        }
-
-        std::set<int> returned_set;
-        for (int i = 0; i < original_set.size(); i++) {
-            returned_set.insert(tc.draw(gs::values_consumed(pool)));
-        }
-
-        assert(original_set == returned_set);
-    });
-}
-
-TEST(Pools, PoolsNoConsume) {
-    hegel::test([](hegel::TestCase& tc) {
-        gs::Pool<int> pool = gs::Pool<int>(tc);
-        uint8_t sz = tc.draw(gs::integers<uint8_t>());
-        std::set<int> original_set =
-            tc.draw(gs::sets(gs::integers<int>(), {.max_size = sz}));
-
-        for (int num : original_set) {
-            pool.add(num);
-        }
-
-        for (int i = 0; i < original_set.size(); i++) {
-            tc.draw(gs::values_reusable(pool));
-        }
-
-        assert(pool.size() == original_set.size());
-    });
-}
 // arrays<T, N>() draws exactly N elements, each from the given generator.
 TEST(Arrays, DrawsExactlyNInRange) {
     hegel::test(
